refactor(observability): move telemetry and event-listener implementations out of core
Core now owns only the neutral observability contracts: the EventListener
trait, event types, and the vendor-neutral gen-AI span conventions in
telemetry. Everything that wires up a vendor or subscriber moved to
everruns-observability:

- OpenTelemetry initialization (OTLP exporter wiring, tracing-subscriber
  layers, TelemetryConfig/TelemetryGuard/init_telemetry) moved to
  crates/observability/src/telemetry.rs.
- CompositeEventListener (fan-out with panic isolation) moved to
  crates/observability/src/composite.rs together with its isolation,
  failure-containment, and ordering tests.
- Removed core's telemetry feature and its opentelemetry, opentelemetry_sdk,
  opentelemetry-otlp, tracing-opentelemetry, and tracing-subscriber
  dependencies; exporter deps are declared only by crates/observability.
- Server and worker binaries install telemetry explicitly through
  everruns-observability (worker gains the dependency).
- Tool narration ownership recorded explicitly: narration is an
  execution-semantic value authored by tool/capability contracts in core;
  exporters only forward the narration field already carried by events.
- New architecture guard scripts/lib/check-observability-isolation.sh
  (pre-push step 14 + CI job) enforces the boundary with source and
  cargo-tree checks, keeping Framework/provider dependency trees free of
  the OTLP/exporter subtree so default Framework builds stay offline.
- Public configuration is unchanged (same OTEL_* env vars); knowledge
  bundle updated (observability, code-organization, tool-narration,
  architecture, log).

Fixes EVE-876
