+++ crates/core/Cargo.toml
@@ -37,7 +37,7 @@ thiserror.workspace = true
 async-trait.workspace = true
 
 # Async runtime
-# Note: "rt" is needed for tokio::spawn in event listener isolation
+# Note: "rt" is needed for tokio::spawn (background tools, parallel act atom)
 tokio = { workspace = true, features = ["sync", "rt", "process", "time", "net"] }
 # `CancellationToken` for cooperative tool-call cancellation (ToolContext).
 tokio-util = { version = "0.7", default-features = false }
@@ -58,19 +58,11 @@ rand.workspace = true
 # Cron parsing for session-schedule minimum-interval enforcement
 cron = "0.17"
 
-# Tracing. `tracing` instrumentation is unconditional; the `tracing-subscriber`
-# registry/fmt layers are only used by telemetry init, so the dep is optional
-# and pulled in by the `telemetry` feature.
+# Tracing. Plain `tracing` instrumentation is unconditional. OpenTelemetry
+# initialization (SDK, OTLP exporter, tracing-subscriber layers) lives in
+# `everruns-observability` (EVE-876); core carries no exporter dependencies.
+# Guard: scripts/lib/check-observability-isolation.sh.
 tracing.workspace = true
-tracing-subscriber = { workspace = true, optional = true }
-
-# OpenTelemetry (optional; behind the `telemetry` feature). The OTLP exporter
-# pulls a large dependency subtree, so provider crates that opt out of core
-# defaults skip it. Plain `tracing` instrumentation stays unconditional.
-opentelemetry = { workspace = true, optional = true }
-opentelemetry_sdk = { workspace = true, optional = true }
-opentelemetry-otlp = { workspace = true, optional = true }
-tracing-opentelemetry = { workspace = true, optional = true }
 
 # Encoding
 base64.workspace = true
@@ -154,15 +146,9 @@ include_dir = { version = "0.7", optional = true }
 # Heavy subtrees stay ON by default so the application crates (server, worker,
 # cli, runtime) and core's own test suite are unchanged. Provider crates depend
 # on core with `default-features = false` to shed them — see crates/openrouter.
-default = ["llm-tests", "telemetry", "a2a", "web-fetch"]
-# OTLP telemetry exporter (OpenTelemetry SDK + tracing-opentelemetry bridge).
-telemetry = [
-    "dep:opentelemetry",
-    "dep:opentelemetry_sdk",
-    "dep:opentelemetry-otlp",
-    "dep:tracing-opentelemetry",
-    "dep:tracing-subscriber",
-]
+# Note: OTLP telemetry moved to `everruns-observability` (EVE-876); no core
+# build — default or otherwise — carries exporter dependencies.
+default = ["llm-tests", "a2a", "web-fetch"]
 # Outbound A2A agent delegation (tonic/prost gRPC stack).
 a2a = ["dep:a2a", "dep:a2a-client"]
 # WebFetch capability (fetchkit HTTP pipeline + ed25519 bot-auth).
