+++ crates/core/src/telemetry.rs
@@ -1,28 +1,14 @@
-// OpenTelemetry Telemetry Module
+// Telemetry Conventions
 //
-// This module provides OpenTelemetry integration for Everruns, including:
-// - Gen-AI semantic conventions for LLM operations
-// - Initialization helpers for OTLP exporters
-// - Span creation helpers with proper attribute naming
-
-// OTLP exporter wiring lives behind the `telemetry` feature so provider crates
-// that only need the gen-AI span conventions (below) do not pull the
-// OpenTelemetry SDK + exporter dependency tree. See crates/core/Cargo.toml.
-#[cfg(feature = "telemetry")]
-use opentelemetry::KeyValue;
-#[cfg(feature = "telemetry")]
-use opentelemetry::trace::TracerProvider as _;
-#[cfg(feature = "telemetry")]
-use opentelemetry_otlp::{SpanExporter, WithExportConfig};
-#[cfg(feature = "telemetry")]
-use opentelemetry_sdk::{
-    Resource,
-    trace::{RandomIdGenerator, Sampler, SdkTracerProvider},
-};
-#[cfg(feature = "telemetry")]
-use std::time::Duration;
-#[cfg(feature = "telemetry")]
-use tracing_subscriber::{EnvFilter, Layer, layer::SubscriberExt, util::SubscriberInitExt};
+// Vendor-neutral gen-AI span metadata contracts shared by execution and
+// exporters:
+// - Gen-AI semantic convention attribute names for LLM operations
+// - Span-name helpers with proper attribute naming
+//
+// OpenTelemetry initialization (OTLP exporter wiring, tracing-subscriber
+// layers, TelemetryConfig/TelemetryGuard) lives in `everruns-observability`
+// (EVE-876) so core carries no OTel SDK, exporter, or subscriber dependencies.
+// Guard: scripts/lib/check-observability-isolation.sh.
 
 // ============================================================================
 // Gen-AI Semantic Conventions
@@ -173,100 +159,8 @@ pub mod gen_ai {
 }
 
 // ============================================================================
-// Telemetry Configuration
+// Crypto Provider
 // ============================================================================
-
-/// Configuration for OpenTelemetry
-#[derive(Debug, Clone)]
-pub struct TelemetryConfig {
-    /// Service name for traces
-    pub service_name: String,
-    /// Service version
-    pub service_version: Option<String>,
-    /// OTLP endpoint (e.g., "http://localhost:4317")
-    pub otlp_endpoint: Option<String>,
-    /// Environment (e.g., "development", "production")
-    pub environment: Option<String>,
-    /// Whether to enable console logging
-    pub enable_console: bool,
-    /// Log filter (e.g., "info", "debug", "everruns=debug")
-    pub log_filter: Option<String>,
-    /// Whether to enable content recording (input/output messages)
-    /// Disabled by default for privacy/security
-    pub record_content: bool,
-}
-
-impl Default for TelemetryConfig {
-    fn default() -> Self {
-        Self {
-            service_name: "everruns".to_string(),
-            service_version: None,
-            otlp_endpoint: None,
-            environment: None,
-            enable_console: true,
-            log_filter: None,
-            record_content: false,
-        }
-    }
-}
-
-impl TelemetryConfig {
-    /// Create configuration from environment variables
-    ///
-    /// Environment variables:
-    /// - `OTEL_SDK_DISABLED`: If "true", disables OpenTelemetry tracing entirely
-    /// - `OTEL_SERVICE_NAME`: Service name (default: "everruns")
-    /// - `OTEL_SERVICE_VERSION`: Service version
-    /// - `OTEL_EXPORTER_OTLP_ENDPOINT`: OTLP endpoint (e.g., "http://localhost:4317")
-    /// - `OTEL_ENVIRONMENT`: Deployment environment
-    /// - `RUST_LOG` or `LOG_LEVEL`: Log filter
-    /// - `OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT`: Record input/output content (standard OTel)
-    /// - `OTEL_RECORD_CONTENT`: Legacy alias for content recording
-    pub fn from_env() -> Self {
-        use crate::config::{env_bool, env_opt_string, env_string};
-
-        let sdk_disabled = env_bool("OTEL_SDK_DISABLED", false);
-
-        Self {
-            service_name: env_string("OTEL_SERVICE_NAME", "everruns"),
-            service_version: env_opt_string("OTEL_SERVICE_VERSION"),
-            otlp_endpoint: if sdk_disabled {
-                None
-            } else {
-                env_opt_string("OTEL_EXPORTER_OTLP_ENDPOINT")
-            },
-            environment: env_opt_string("OTEL_ENVIRONMENT"),
-            enable_console: true,
-            log_filter: env_opt_string("RUST_LOG").or_else(|| env_opt_string("LOG_LEVEL")),
-            // Standard OTel env var, with legacy alias fallback
-            record_content: std::env::var("OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT")
-                .or_else(|_| std::env::var("OTEL_RECORD_CONTENT"))
-                .map(|v| v.to_lowercase() == "true")
-                .unwrap_or(false),
-        }
-    }
-}
-
-// ============================================================================
-// Initialization
-// ============================================================================
-
-/// Guard that shuts down the tracer provider when dropped
-#[cfg(feature = "telemetry")]
-pub struct TelemetryGuard {
-    _provider: Option<SdkTracerProvider>,
-}
-
-#[cfg(feature = "telemetry")]
-impl Drop for TelemetryGuard {
-    fn drop(&mut self) {
-        if let Some(provider) = self._provider.take()
-            && let Err(e) = provider.shutdown()
-        {
-            eprintln!("Failed to shutdown tracer provider: {:?}", e);
-        }
-    }
-}
 
 /// Install the rustls crypto provider (ring) for TLS.
 ///
@@ -278,122 +172,6 @@ impl Drop for TelemetryGuard {
 pub fn install_crypto_provider() {
     // Already-installed is expected if called from multiple init paths
     let _ = rustls::crypto::ring::default_provider().install_default();
-}
-
-/// Initialize OpenTelemetry with the given configuration
-///
-/// Returns a guard that will shut down the tracer provider when dropped.
-/// Keep this guard alive for the lifetime of your application.
-///
-/// # Example
-///
-/// ```ignore
-/// use everruns_core::telemetry::{init_telemetry, TelemetryConfig};
-///
-/// #[tokio::main]
-/// async fn main() {
-///     let config = TelemetryConfig::from_env();
-///     let _guard = init_telemetry(config);
-///     // ... your application code
-/// }
-/// ```
-#[cfg(feature = "telemetry")]
-pub fn init_telemetry(config: TelemetryConfig) -> TelemetryGuard {
-    install_crypto_provider();
-
-    // Build resource with service info
-    let mut resource_attrs = vec![KeyValue::new("service.name", config.service_name.clone())];
-
-    if let Some(version) = &config.service_version {
-        resource_attrs.push(KeyValue::new("service.version", version.clone()));
-    }
-
-    if let Some(env) = &config.environment {
-        resource_attrs.push(KeyValue::new("deployment.environment", env.clone()));
-    }
-
-    let resource = Resource::builder().with_attributes(resource_attrs).build();
-
-    // Build log filter
-    let filter = config
-        .log_filter
-        .as_ref()
-        .and_then(|f| EnvFilter::try_new(f).ok())
-        .unwrap_or_else(|| EnvFilter::new("info"));
-
-    // Build console layer if enabled
-    let console_layer = if config.enable_console {
-        Some(
-            tracing_subscriber::fmt::layer()
-                .with_target(true)
-                .with_filter(filter),
-        )
-    } else {
-        None
-    };
-
-    // Build OTLP tracer if endpoint is configured
-    let (tracer_provider, otel_layer, otel_status) = if let Some(endpoint) = &config.otlp_endpoint {
-        match build_otlp_tracer(endpoint, resource) {
-            Ok((provider, tracer)) => {
-                let layer = tracing_opentelemetry::layer().with_tracer(tracer);
-                (Some(provider), Some(layer), Some(Ok(endpoint.clone())))
-            }
-            Err(e) => (None, None, Some(Err(e.to_string()))),
-        }
-    } else {
-        (None, None, None)
-    };
-
-    // Initialize the subscriber
-    tracing_subscriber::registry()
-        .with(console_layer)
-        .with(otel_layer)
-        .init();
-
-    // Log OTEL status after subscriber is initialized
-    match otel_status {
-        Some(Ok(endpoint)) => {
-            tracing::info!(endpoint = %endpoint, "OpenTelemetry tracing enabled");
-        }
-        Some(Err(e)) => {
-            tracing::warn!(error = %e, "Failed to initialize OTLP tracer, continuing without tracing");
-        }
-        None => {
-            tracing::debug!("OpenTelemetry tracing disabled: OTEL_EXPORTER_OTLP_ENDPOINT not set");
-        }
-    }
-
-    TelemetryGuard {
-        _provider: tracer_provider,
-    }
-}
-
-#[cfg(feature = "telemetry")]
-fn build_otlp_tracer(
-    endpoint: &str,
-    resource: Resource,
-) -> Result<
-    (SdkTracerProvider, opentelemetry_sdk::trace::Tracer),
-    Box<dyn std::error::Error + Send + Sync>,
-> {
-    // Use HTTP OTLP instead of gRPC - more reliable with Docker DNS
-    let exporter = SpanExporter::builder()
-        .with_http()
-        .with_endpoint(endpoint)
-        .with_timeout(Duration::from_secs(10))
-        .build()?;
-
-    let provider = SdkTracerProvider::builder()
-        .with_batch_exporter(exporter)
-        .with_sampler(Sampler::AlwaysOn)
-        .with_id_generator(RandomIdGenerator::default())
-        .with_resource(resource)
-        .build();
-
-    let tracer = provider.tracer("everruns");
-
-    Ok((provider, tracer))
 }
 
 // ============================================================================
@@ -494,14 +272,5 @@ mod tests {
             embeddings_span_name("text-embedding-ada-002"),
             "embeddings text-embedding-ada-002"
         );
-    }
-
-    #[test]
-    fn test_config_defaults() {
-        let config = TelemetryConfig::default();
-        assert_eq!(config.service_name, "everruns");
-        assert!(config.otlp_endpoint.is_none());
-        assert!(config.enable_console);
-        assert!(!config.record_content);
     }
 }
