+++ crates/worker/src/main.rs
@@ -2,7 +2,7 @@
 // Decision: Uses WorkerAppBuilder for composable worker setup
 
 use anyhow::Result;
-use everruns_core::telemetry::{TelemetryConfig, init_telemetry};
+use everruns_observability::{TelemetryConfig, init_telemetry};
 use everruns_worker::{TaskWorkerConfig, WorkerAppBuilder};
 
 // Use mimalloc as the global allocator. The worker runs heavy concurrent
