+++ knowledge/foundations/architecture.md
@@ -550,7 +550,7 @@ Observability is decoupled from business logic through the `EventListener` trait
 
 **Key components**:
 - `EventListener` trait — Interface for observability backends
-- `OtelEventListener` (`core/src/otel_listener.rs`) — Generates OTel spans from events
+- `OtelEventListener` (`crates/observability/src/otel.rs`) — Generates OTel spans from events
 - `EventService` (`server/src/services/event.rs`) — Notifies listeners after event persistence
 
 **Event-to-span mapping** (following gen-ai semantic conventions):
