+++ scripts/lib/pre-push.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 # Pre-push checks: fast local validation to catch CI failures early (~30s).
-# Runs formatting, linting, lockfile, migration, test/example enumeration, knowledge, capability-contract, test-support-isolation, and attribution checks.
+# Runs formatting, linting, lockfile, migration, test/example enumeration, knowledge, capability-contract, test-support-isolation, observability-isolation, and attribution checks.
 # Usage: just pre-push (or: bash scripts/lib/pre-push.sh)
 
 source "$(dirname "${BASH_SOURCE[0]}")/common.sh"
@@ -13,7 +13,7 @@ echo "🔒 Running pre-push checks..."
 echo ""
 
 # 1. Rust formatting
-echo "1/14 Rust formatting"
+echo "1/15 Rust formatting"
 if cargo fmt --check 2>/dev/null; then
   pass "cargo fmt"
 else
@@ -21,7 +21,7 @@ else
 fi
 
 # 2. Clippy
-echo "2/14 Rust linting"
+echo "2/15 Rust linting"
 if cargo clippy --all-targets --all-features -- -D warnings 2>/dev/null; then
   pass "clippy"
 else
@@ -29,7 +29,7 @@ else
 fi
 
 # 3. Cargo.lock freshness
-echo "3/14 Cargo.lock freshness"
+echo "3/15 Cargo.lock freshness"
 if cargo fetch --locked 2>/dev/null; then
   pass "Cargo.lock up to date"
 else
@@ -37,7 +37,7 @@ else
 fi
 
 # 4. UI formatting (skip if node_modules missing)
-echo "4/14 UI formatting"
+echo "4/15 UI formatting"
 if [ -d "$PROJECT_ROOT/apps/ui/node_modules" ]; then
   if (cd "$PROJECT_ROOT/apps/ui" && pnpm run format:check 2>/dev/null); then
     pass "UI format"
@@ -49,7 +49,7 @@ else
 fi
 
 # 5. UI linting (skip if node_modules missing)
-echo "5/14 UI linting"
+echo "5/15 UI linting"
 if [ -d "$PROJECT_ROOT/apps/ui/node_modules" ]; then
   if (cd "$PROJECT_ROOT/apps/ui" && pnpm run lint 2>/dev/null); then
     pass "UI lint"
@@ -61,7 +61,7 @@ else
 fi
 
 # 6. Migration ordering check
-echo "6/14 Migration ordering"
+echo "6/15 Migration ordering"
 if MIGRATION_ORDERING_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/lib/check-migration-ordering.sh" 2>&1
 )"; then
@@ -72,7 +72,7 @@ else
 fi
 
 # 7. Migration immutability check
-echo "7/14 Migration immutability"
+echo "7/15 Migration immutability"
 if MIGRATION_IMMUTABILITY_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/lib/check-migration-immutability.sh" 2>&1
 )"; then
@@ -83,7 +83,7 @@ else
 fi
 
 # 8. Server integration test enumeration
-echo "8/14 Server test enumeration"
+echo "8/15 Server test enumeration"
 if SERVER_TEST_ENUM_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/lib/check-server-test-enumeration.sh" 2>&1
 )"; then
@@ -94,7 +94,7 @@ else
 fi
 
 # 9. Public everruns example inventory and compile check
-echo "9/14 Public everruns examples"
+echo "9/15 Public everruns examples"
 if EVERRUNS_EXAMPLES_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/lib/check-everruns-examples.sh" 2>&1
 )"; then
@@ -105,7 +105,7 @@ else
 fi
 
 # 10. Knowledge bundle conformance
-echo "10/14 Knowledge bundle conformance"
+echo "10/15 Knowledge bundle conformance"
 if KNOWLEDGE_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/test-knowledge-okf.sh" 2>&1
 )"; then
@@ -116,7 +116,7 @@ else
 fi
 
 # 11. Published crate documentation contract
-echo "11/14 Published crate documentation"
+echo "11/15 Published crate documentation"
 if PUBLISHED_DOCS_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/test-published-crate-docs.sh" 2>&1
 )"; then
@@ -127,7 +127,7 @@ else
 fi
 
 # 12. Capability contract architecture guard (EVE-873)
-echo "12/14 Capability contract guard"
+echo "12/15 Capability contract guard"
 if CAPABILITY_CONTRACT_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/lib/check-capability-contract.sh" 2>&1
 )"; then
@@ -138,7 +138,7 @@ else
 fi
 
 # 13. Test-support isolation guard (EVE-875)
-echo "13/14 Test-support isolation guard"
+echo "13/15 Test-support isolation guard"
 if TEST_SUPPORT_ISOLATION_OUTPUT="$(
   bash "$PROJECT_ROOT/scripts/lib/check-test-support-isolation.sh" 2>&1
 )"; then
@@ -148,8 +148,19 @@ else
   fail "test-support isolation guard failed"
 fi
 
-# 14. Commit author attribution check
-echo "14/14 Commit author attribution"
+# 14. Observability isolation guard (EVE-876)
+echo "14/15 Observability isolation guard"
+if OBSERVABILITY_ISOLATION_OUTPUT="$(
+  bash "$PROJECT_ROOT/scripts/lib/check-observability-isolation.sh" 2>&1
+)"; then
+  pass "$OBSERVABILITY_ISOLATION_OUTPUT"
+else
+  printf '%s\n' "$OBSERVABILITY_ISOLATION_OUTPUT" | sed 's/^/   /'
+  fail "observability isolation guard failed"
+fi
+
+# 15. Commit author attribution check
+echo "15/15 Commit author attribution"
 if ! resolve_commit_git_identity; then
   fail "commit identity invalid — fix git config or set GIT_USER_NAME/GIT_USER_EMAIL to a real user"
 elif OFFENDING_COMMIT="$(find_agent_like_outgoing_commit)"; then
